test: cli of import-grammars #55
